pyproject.toml
[project]
name = "automated-news-livestream"
version = "0.2.0"
description = "Automated News Livestream Intelligence System - CPU-only multimodal news monitoring with VLM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "patchright>=1.47.0",
    "opencv-python>=4.10.0",
    "pillow>=10.4.0",
    "numpy>=1.26.4",
    "pyyaml>=6.0.2",
    "python-dateutil>=2.9.0",
    "llama-cpp-python>=0.2.90",
    "pywhispercpp>=1.2.0",
    "pytesseract>=0.3.10",
    "ultralytics>=8.0.0",
    "easyocr>=1.7.0",
    "requests>=2.32.0",
    "tqdm>=4.66.0",
    "transformers>=4.45.0",
    # NOTE: torch must be installed separately for CPU-only
    # Run: uv pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
    "accelerate>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "black>=24.8.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# PyTorch CPU-only installation
# Use this command instead of regular uv sync:
# uv pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# Then: uv pip install -e .

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []
